server and client part: put new rating to deviceTable in DB

+++ server/models/models.js
@@ -19,7 +19,7 @@ const Device = sequelize.define('device', {
     id:{type:DataTypes.INTEGER, primaryKey: true, autoIncrement: true},
     name:{type:DataTypes.STRING, unique: true, allowNull: false},
     price:{type:DataTypes.INTEGER, allowNull: false},
-    rating:{type:DataTypes.INTEGER, defaultValue: null},
+    rating:{type:DataTypes.FLOAT, defaultValue: null},
     img:{type:DataTypes.STRING, allowNull: false}
 })
